systemz: fix #1062

+++ arch/SystemZ/SystemZInstPrinter.c
@@ -302,7 +302,7 @@ static void printPCRelOperand(MCInst *MI, int OpNum, SStream *O)
 				SStream_concat(O, "%u", imm);
 		} else {
 			if (imm < -HEX_THRESHOLD)
-				SStream_concat(O, "-0x%x", -imm);
+				SStream_concat(O, "-0x%x", (unsigned int)-imm);
 			else
 				SStream_concat(O, "-%u", -imm);
 		}
